+++ src/cpp/include/openvino/genai/visual_language/vlm_pipeline.hpp
