+++ src/cpp/src/visual_language/clip.cpp
