+++ src/cpp/src/visual_language/vlm_config.cpp
